requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
ciso8601==2.1.3
idna==2.9
iso-639==0.4.5
iso3166==1.0.1
iso8601==0.1.12
isodate==0.6.0
kafka-python==2.0.1
numpy==1.18.4
opencv-python==4.2.0.34
Pillow==7.1.2
pycryptodome==3.9.7
PySocks==1.7.1
pytesseract==0.3.4
redis==3.5.1
requests==2.23.0
six==1.14.0
streamlink==1.4.1
tesserocr==2.5.1
urllib3==1.25.9
websocket-client==0.57.0
